requirements.txt
django~=4.1.2
djangorestframework~=3.14.0
django-celery-beat==2.4.0